scroll to input enhanced
one thing that is a major upgrade in this version is you don't actually have to touch the inputs. You can programmatically focus them (which I do constantly in `onSubmitEditing` where I automatically focus the next input). 

In some regards it has less features there. It's very simplistic. It simply scrolls the input to the top of the scrollview if it's behind the keyboard. It doesn't attempt to position the input directly above the keyboard. It also doesn't change scrollview `contentInsets` etc. It's just trying to be a stable base version. But the code is straightforward enough that anyone could add the feature to scroll inputs directly above the keyboard rather than all the way to the top--I just didn't want to get into varying heights that textareas might have and the problems that might bring.